demo_06/requirements.txt
google-cloud-bigquery==1.8.1
google-cloud-core==0.29.1